pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "newsapi-python>=0.2.7",
    "openai>=1.64.0",
    "pandas>=2.2.3",
    "streamlit>=1.42.2",
]
